# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Portal Berita",
        "description": "Portal Berita API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local server"
        },
        {
            "url": "http://staging.example.com",
            "description": "Staging server"
        },
        {
            "url": "http://example.com",
            "description": "Production server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/pelanggan": {
            "get": {
                "tags": [
                    "Pelanggan"
                ],
                "summary": "List or search pelanggan by name",
                "description": "Retrieve all pelanggan or filter pelanggan by partial match on 'nama' using query string",
                "operationId": "listOrSearchPelanggan",
                "parameters": [
                    {
                        "name": "nama",
                        "in": "query",
                        "description": "Optional. Partial or full name of the pelanggan to search for",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Budi"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of pelanggan or search result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pelanggan retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nama": {
                                                        "type": "string",
                                                        "example": "Budi"
                                                    },
                                                    "kontak": {
                                                        "type": "string",
                                                        "example": "08123456789"
                                                    },
                                                    "alamat": {
                                                        "type": "string",
                                                        "example": "Jl. Merdeka 10"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No matching pelanggan found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No matching pelanggan found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pelanggan"
                ],
                "summary": "Create a new pelanggan",
                "operationId": "createPelanggan",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "kontak",
                                    "alamat"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Siti"
                                    },
                                    "kontak": {
                                        "type": "string",
                                        "example": "081298765432"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Diponegoro 15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pelanggan created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Pelanggan created successfully",
                                    "data": {
                                        "id": 2,
                                        "nama": "Siti",
                                        "kontak": "081298765432",
                                        "alamat": "Jl. Diponegoro 15"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pelanggan/{id}": {
            "get": {
                "tags": [
                    "Pelanggan"
                ],
                "summary": "Get pelanggan by ID",
                "operationId": "getPelangganById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pelanggan retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Pelanggan retrieved successfully",
                                    "data": {
                                        "id": 1,
                                        "nama": "Budi",
                                        "kontak": "08123456789",
                                        "alamat": "Jl. Merdeka 10"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pelanggan not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Pelanggan not found"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Pelanggan"
                ],
                "summary": "Update existing pelanggan",
                "operationId": "updatePelanggan",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Andi"
                                    },
                                    "kontak": {
                                        "type": "string",
                                        "example": "081234000000"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Sudirman 23"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pelanggan updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Pelanggan updated successfully",
                                    "data": {
                                        "id": 1,
                                        "nama": "Andi",
                                        "kontak": "081234000000",
                                        "alamat": "Jl. Sudirman 23"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pelanggan"
                ],
                "summary": "Delete a pelanggan",
                "operationId": "deletePelanggan",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pelanggan deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Pelanggan deleted successfully"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}